LLM/Image_text_search/requirements.txt
torch==2.2.2
torchvision==0.17.2
faiss-cpu
clip-by-openai
ftfy
regex
tqdm
numpy
Pillow
streamlit
